Remove deprecated backend files including client, config, llm_layer, log_fetcher, logging_config, regex_parsers, server, test_server, topics, and vector_db. This cleanup streamlines the codebase by eliminating unused components.
