+++ backend/config/__init__.py
@@ -1,0 +1,13 @@
+"""
+IntelliOS Config Module
+
+This module contains configuration components:
+- Application settings and environment variables
+- Logging configuration
+- Pydantic schemas for data validation
+"""
+
+from .config import ActivityLog
+from .logging_config import setup_logging
+
+__all__ = ['ActivityLog', 'setup_logging']
