+++ backend/core/__init__.py
@@ -1,0 +1,13 @@
+"""
+IntelliOS Core Module
+
+This module contains core system components:
+- Digital DNA state management
+- Topic definitions and categorization
+- LangGraph workflow orchestration
+- State persistence and retrieval
+"""
+
+from .topics import TOPICS, TOPIC_EXAMPLES
+
+__all__ = ['TOPICS', 'TOPIC_EXAMPLES']
