+++ backend/pipeline/__init__.py
@@ -1,0 +1,22 @@
+"""
+IntelliOS Pipeline Module
+
+This module contains the data processing pipeline components:
+- Log fetching from Windows Event Logs
+- Deterministic parsing using regex patterns
+- LLM-based parsing for complex logs
+- Log processing orchestration
+"""
+
+from .log_processor import process_logs, parse_arguments
+from .log_fetcher import fetch_windows_event_logs
+from .regex_parsers import parse_with_regex
+from .llm_layer import parse_with_llm
+
+__all__ = [
+    'process_logs',
+    'parse_arguments', 
+    'fetch_windows_event_logs',
+    'parse_with_regex',
+    'parse_with_llm'
+]
