+++ backend/utils/__init__.py
@@ -1,0 +1,14 @@
+"""
+IntelliOS Utils Module
+
+This module contains utility components:
+- API client for external communication
+- Testing utilities
+- Common helper functions
+- Custom decorators
+"""
+
+from .client import IntelliOSClient
+from .test_server import test_api_endpoints
+
+__all__ = ['IntelliOSClient', 'test_api_endpoints']
